+++ actix-files/CHANGES.md
@@ -1,7 +1,9 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+- Added `Files::try_compressed()` to support serving pre-compressed static files [#2615]
 
+[#2615]: https://github.com/actix/actix-web/pull/2615
 
 ## 0.6.0-beta.15 - 2022-01-21
 - No significant changes since `0.6.0-beta.14`.